:sparkles: Decoupled GraphHandler with GraphHandlerForKG and GraphHandlerForDF classes
