+++ code/load/mlentory_load/core/__init__.py
@@ -1,10 +1,14 @@
 from .Entities import Model, HFModel
 from .GraphHandler import GraphHandler
+from .GraphHandlerForKG import GraphHandlerForKG
+from .GraphHandlerForDF import GraphHandlerForDF
 from .LoadProcessor import LoadProcessor
 
 __all__ = [
     "Model",
     "HFModel",
     "GraphHandler",
+    "GraphHandlerForKG",
+    "GraphHandlerForDF",
     "LoadProcessor",
 ]
